refactor(hub): extract backup/dump/load from vault.ts (Phase 5 A4)
